Move RemoteLogForwarder to server
