+++ endermux-server/build.gradle.kts
@@ -12,6 +12,8 @@ indra {
 
 dependencies {
   api(project(":endermux-common"))
+  compileOnly(project(":endermux-log4j-plugins"))
+  compileOnly(libs.log4jCore)
   implementation(libs.slf4jApi)
   compileOnlyApi(libs.jspecify)
   compileOnly(libs.adventureApi)
